#42 - Add encoder for form data transfer testing
Form data should be serialized in a different format from APIs.
Create a utility for converting objects into strings in the form data format for testing purposes, and write tests for it.
If you don't have a utility, you can also use 'post().param()' to input values, but it doesn't seem to match the actual POST request transmission, and it's inconvenient because you have to input parameters instead of objects.