#80 - Design child-comment domain
Add code to establish parent-child relationships within the child-comment domain.
Apply cascading rules to reflect changes in the child comments collection in the query.
Set up a unidirectional association by expressing the parent comment as a 'Long' id instead of an entity.
Add a method to allow the addition of child comments.